Avoid deadlock, demonstrate asynch communication.
Depending on the MPI implementation, the `MPI_Send` in `heat.c` could
lead to deadlock, as every process would block before any process hits
an `MPI_Recv`. To address this, the call was changed to an `MPI_Isend`.

Since this required two asynchronous calls anyway, the `MPI_Recv` was
moved to demonstrate interleaving of communication and computation (i.e.
the internal ghost cells are computed after the `MPI_Isend` is posted,
but before the `MPI_Recv`.